requirements.txt
PyYAML==3.11
argparse==1.2.1
configobj==5.0.6
coverage==4.2
coveralls==1.0
crontab==0.20.3
docopt==0.6.2
funcsigs==1.0.2
gendobot==2.2.1
mock==1.3.0
nose==1.3.6
pbr==1.10.0
pyfieri==1.2.2
requests==2.12.0
six==1.10.0
slackclient==0.16
websocket-client==0.34.0
wheel==0.24.0
wsgiref==0.1.2
